+ obj interactif avec joueur et enemy pour attack et heal

+++ Assets/Features/Objet/ItemManager.cs
@@ -137,7 +137,7 @@ public class ItemManager : MonoBehaviour
             item.TriggerVisualUpdate();
         }
     }
-    public void ChangeItemEffect(List<ItemBrain> items, float Effect)
+    public void ChangeItemEffect(List<ItemBrain> items, int Effect)
     {
         foreach (ItemBrain item in items)
         {
